update structure for Market Organizer Project
