+++ MarketOrganizer.Api/.gitignore
@@ -342,3 +342,5 @@ healthchecksdb
 
 # Backup folder for Package Reference Convert tool in Visual Studio 2017
 MigrationBackup/
+
+appsettings.Development.json
