+++ MarketOrganizer.Api/Controllers/ItemsController.cs
@@ -1,11 +1,11 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using api.Models;
+using MarketOrganizer.Data.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
-namespace api.Controllers
+namespace MarketOrganizer.Api.Controllers
 {
   [Route("api/[controller]")]
   [ApiController]
