+++ MarketOrganizer.Api/Properties/launchSettings.json
