+++ MarketOrganizer.Api/appsettings.json
