+++ MarketOrganizer.Data/Models/ItemsContext.cs
@@ -1,0 +1,11 @@
+using Microsoft.EntityFrameworkCore;
+
+namespace MarketOrganizer.Data.Models
+{
+  public class ItemsContext : DbContext
+  {
+    public ItemsContext(DbContextOptions<ItemsContext> options) : base(options) { }
+
+    public DbSet<Items> Item { get; set; }
+  }
+}
