+++ MarketOrganizer.sln
@@ -1,0 +1,48 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26124.0
+MinimumVisualStudioVersion = 15.0.26124.0
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MarketOrganizer.Api", "MarketOrganizer.Api\MarketOrganizer.Api.csproj", "{E8CBD0CA-F1BA-4745-BA3F-321B27D85EA2}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MarketOrganizer.Data", "MarketOrganizer.Data\MarketOrganizer.Data.csproj", "{57715E58-590A-4250-9BBE-B63669AB7200}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E8CBD0CA-F1BA-4745-BA3F-321B27D85EA2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E8CBD0CA-F1BA-4745-BA3F-321B27D85EA2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E8CBD0CA-F1BA-4745-BA3F-321B27D85EA2}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{E8CBD0CA-F1BA-4745-BA3F-321B27D85EA2}.Debug|x64.Build.0 = Debug|Any CPU
+		{E8CBD0CA-F1BA-4745-BA3F-321B27D85EA2}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{E8CBD0CA-F1BA-4745-BA3F-321B27D85EA2}.Debug|x86.Build.0 = Debug|Any CPU
+		{E8CBD0CA-F1BA-4745-BA3F-321B27D85EA2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E8CBD0CA-F1BA-4745-BA3F-321B27D85EA2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E8CBD0CA-F1BA-4745-BA3F-321B27D85EA2}.Release|x64.ActiveCfg = Release|Any CPU
+		{E8CBD0CA-F1BA-4745-BA3F-321B27D85EA2}.Release|x64.Build.0 = Release|Any CPU
+		{E8CBD0CA-F1BA-4745-BA3F-321B27D85EA2}.Release|x86.ActiveCfg = Release|Any CPU
+		{E8CBD0CA-F1BA-4745-BA3F-321B27D85EA2}.Release|x86.Build.0 = Release|Any CPU
+		{57715E58-590A-4250-9BBE-B63669AB7200}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{57715E58-590A-4250-9BBE-B63669AB7200}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{57715E58-590A-4250-9BBE-B63669AB7200}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{57715E58-590A-4250-9BBE-B63669AB7200}.Debug|x64.Build.0 = Debug|Any CPU
+		{57715E58-590A-4250-9BBE-B63669AB7200}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{57715E58-590A-4250-9BBE-B63669AB7200}.Debug|x86.Build.0 = Debug|Any CPU
+		{57715E58-590A-4250-9BBE-B63669AB7200}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{57715E58-590A-4250-9BBE-B63669AB7200}.Release|Any CPU.Build.0 = Release|Any CPU
+		{57715E58-590A-4250-9BBE-B63669AB7200}.Release|x64.ActiveCfg = Release|Any CPU
+		{57715E58-590A-4250-9BBE-B63669AB7200}.Release|x64.Build.0 = Release|Any CPU
+		{57715E58-590A-4250-9BBE-B63669AB7200}.Release|x86.ActiveCfg = Release|Any CPU
+		{57715E58-590A-4250-9BBE-B63669AB7200}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
