+++ global.json
@@ -1,0 +1,5 @@
+{
+  "sdk": {
+    "version": "2.1.504"
+  }
+}
